requirements.txt
sentence-transformers==2.7.0
faiss-cpu==1.8.0.post1
pypdf==4.3.1
numpy==1.26.4
tqdm==4.66.4
python-dotenv==1.0.1
typer==0.12.3
click==8.1.7
rich==13.7.1
openai==1.40.0
python-dotenv==1.0.1
transformers>=4.41.0
torch>=2.0.0
sentencepiece
sentence-transformers
fastapi
uvicorn[standard]
python-multipart
